martes16/01

+++ u4_tiendaLaravel/database/migrations/2024_01_15_084512_create_pedido__productos_table.php
@@ -19,7 +19,7 @@ return new class extends Migration
             //FK convencion de nombres Laravel
             $table->foreignId('pedido_id')->constrained()->onDelete('restrict')->onUpdate('cascade');
             //FK SIN convencion de nombres Laravel
-            $table->integer('producto');
+            $table->unsignedBigInteger('producto');
             $table->foreign('producto')->references('id')->on('productos')->onDelete('restrict')->onUpdate('cascade');
         });
     }
